requirements.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

django
django-crispy-forms
django-model-utils
psycopg2
uwsgi
